requirements.txt
langchain
langchain-community
langchain-ollama
langchain-experimental
neo4j
tiktoken
yfiles_jupyter_graphs
python-dotenv
json-repair
langchain-openai
langchain_core
langchain-google-genai
google-generativeai
torch>=2.0.0
transformers>=4.30.0
networkx>=3.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0

